# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tea Sequence API
  version: 1.0.0
  description: >-
    Stable v1 contract for Tea Sequence. The initial configuration response
    explicitly describes non-production demonstration mode.
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /api/v1/admin/auth/login:
    post:
      operationId: loginAdmin
      parameters: [{$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginCreate'}}}}
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
        '401': {$ref: '#/components/responses/ApiFailure'}
        '403': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/admin/auth/me:
    get:
      operationId: getCurrentAdmin
      security: [{AdminSessionCookie: []}]
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}, '401': {$ref: '#/components/responses/ApiFailure'}}
  /api/v1/admin/auth/logout:
    post:
      operationId: logoutAdmin
      security: [{AdminSessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CsrfToken'}]
      responses: {'204': {description: Session revoked}, '403': {$ref: '#/components/responses/ApiFailure'}}
  /api/v1/admin/auth/password:
    patch:
      operationId: changeAdminPassword
      security: [{AdminSessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CsrfToken'}]
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}}
  /api/v1/admin/users:
    get:
      operationId: listAdminUsers
      security: [{AdminSessionCookie: []}]
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}}
    post:
      operationId: createAdminUser
      security: [{AdminSessionCookie: []}]
      parameters: [{$ref: '#/components/parameters/CsrfToken'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'201': {$ref: '#/components/responses/ApiSuccess'}}
  /api/v1/admin/users/{user_id}:
    patch:
      operationId: patchAdminUser
      security: [{AdminSessionCookie: []}]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/CsrfToken'}
        - {$ref: '#/components/parameters/IfMatch'}
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}, '409': {$ref: '#/components/responses/ApiFailure'}, '428': {$ref: '#/components/responses/ApiFailure'}}
  /api/v1/admin/users/{user_id}/password-reset:
    post:
      operationId: resetAdminPassword
      security: [{AdminSessionCookie: []}]
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/CsrfToken'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}}
...
components:
  securitySchemes:
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: tea_sequence_session
      description: HttpOnly administrator session cookie for later admin operations.
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required by later cookie-authenticated state-changing operations.
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Required by later state-changing operations; must be a UUID.
      schema:
        type: string
        format: uuid
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string, pattern: '^"rv-[1-9][0-9]*"$'}
...
  responses:
    ApiSuccess:
      description: Standard success envelope; operation-specific data is defined by the prose contract.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/SuccessObject'}
    ApiFailure:
      description: Standard failure envelope; clients branch on error.code.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
    SuccessObject:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {type: object}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
    LoginCreate:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 64}
        password: {type: string, minLength: 1, maxLength: 128}
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [request_id, server_time]
      properties:
        request_id:
          type: string
          format: uuid
        server_time:
          type: string
          format: date-time
    ErrorDetail:
      type: object
      additionalProperties: true
      properties:
        field:
          type: string
        reason:
          type: string
    ApiErrorBody:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorBody'
        meta:
          $ref: '#/components/schemas/ResponseMeta'